backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
ultralytics==8.2.18
torch==2.3.0
torchvision==0.18.0
opencv-python==4.9.0.80
pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.1
